requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-compress>=1.14
python-dotenv==1.1.0
chromadb==0.6.3
openai-whisper==20240930
sentence-transformers==3.4.1
requests==2.32.3
gunicorn==23.0.0
psutil==5.9.8
